(feat) Database connected in server; able to create tables and save data
