+++ database/rethinkdb.js
@@ -1,32 +1,33 @@
-/*TODO: Add to Koa server */
+// // /*TODO: Add to Koa server */
 
-import thinky from './thinkylocal';
+// import thinky from './thinkylocal';
 
-// Import Rethink
-import r from 'rethinkdb';
-import config from 'config.js';
+// // Import Rethink
+// import r from 'rethinkdb';
+// import config from 'config.js';
 
-// Create a rethinkdb connection - they save it in req._rdbConn
-app.use(createConnection);
+// // Create a rethinkdb connection, and save it in req._rdbConn
+// app.use(createConnection);
 
-app.use(closeConnection);
+// app.use(closeConnection);
 
 
-function* createConnection(next) {
-  try {
-    const conn = yield r.connect(config.rethinkdb);
-    this._rdbConn = conn;
-  }
-  catch(err) {
-    this.status = 500;
-    this.body = err.message || 'Could not connect to DB';
-  }
-  yield next;
-}
+// function* createConnection(next) {
+//   try {
+//     const conn = yield r.connect(config.rethinkdb);
+//     this._rdbConn = conn;
+//   }
+//   catch(err) {
+//     this.status = 500;
+//     this.body = err.message || http:STATUS_CODES[this.status];
+//   }
+//   yield next;
+// }
 
-/*
- * Close the RethinkDB connection
- */
-function* closeConnection(next) {
-    this._rdbConn.close();
-}
+// /*
+//  * Close the RethinkDB connection
+//  */
+// function* closeConnection(next) {
+//     this._rdbConn.close();
+//      yield next;
+// }
