--- database/rethinkdb_data/log_file
@@ -1,26 +1,0 @@
-2016-03-16T22:51:34.858591000 0.762742s notice: Recursively removing directory /Users/aimeerosato/hackreactor/thesis/numbers-game/database/rethinkdb_data/tmp
-2016-03-16T22:51:34.860226000 0.764380s notice: Initializing directory /Users/aimeerosato/hackreactor/thesis/numbers-game/database/rethinkdb_data
-2016-03-16T22:51:34.860326000 0.764475s info: Creating a default database for your convenience. (This is because you ran 'rethinkdb' without 'create', 'serve', or '--join', and the directory '/Users/aimeerosato/hackreactor/thesis/numbers-game/database/rethinkdb_data' did not already exist or is empty.)
-2016-03-16T22:51:34.860366000 0.764514s notice: Running rethinkdb 2.2.5 (CLANG 7.0.2 (clang-700.1.81))...
-2016-03-16T22:51:34.966814000 0.870965s notice: Running on Darwin 15.3.0 x86_64
-2016-03-16T22:51:34.966878000 0.871027s notice: Loading data from directory /Users/aimeerosato/hackreactor/thesis/numbers-game/database/rethinkdb_data
-2016-03-16T22:51:36.081306000 1.985455s info: Automatically using cache size of 100 MB
-2016-03-16T22:51:36.081330000 1.985478s warn: Cache size does not leave much memory for server and query overhead (available memory: 512 MB).
-2016-03-16T22:51:36.081340000 1.985488s warn: Cache size is very low and may impact performance.
-2016-03-16T22:51:36.095739000 1.999888s notice: Listening for intracluster connections on port 29015
-2016-03-16T22:51:36.122988000 2.027137s notice: Listening for client driver connections on port 28015
-2016-03-16T22:51:36.123167000 2.027316s notice: Listening for administrative HTTP connections on port 8080
-2016-03-16T22:51:36.123169000 2.027317s notice: Listening on addresses: 127.0.0.1, ::1
-2016-03-16T22:51:36.123173000 2.027321s notice: To fully expose RethinkDB on the network, bind to all addresses by running rethinkdb with the `--bind all` command line option.
-2016-03-16T22:51:36.123173001 2.027322s notice: Server ready, "Aimees_MBP_ahc" ccb6eff2-accf-45ce-ad7f-cf57d9d893e4
-2016-03-16T23:10:08.142123000 1114.060575s info: Table ba9c9765-eb04-489f-b692-a821cf98191f: Added replica on this server.
-2016-03-16T23:10:08.236206000 1114.154658s info: Table dafe27fb-8dad-4d42-835f-d5fb75720e0e: Added replica on this server.
-2016-03-16T23:10:09.296533000 1115.214985s info: Table ba9c9765-eb04-489f-b692-a821cf98191f: Starting a new Raft election for term 1.
-2016-03-16T23:10:09.541425000 1115.459877s info: Table ba9c9765-eb04-489f-b692-a821cf98191f: This server is Raft leader for term 1. Latest log index is 0.
-2016-03-16T23:10:10.072931000 1115.991408s info: Table dafe27fb-8dad-4d42-835f-d5fb75720e0e: Starting a new Raft election for term 1.
-2016-03-16T23:10:10.272722000 1116.191200s info: Table dafe27fb-8dad-4d42-835f-d5fb75720e0e: This server is Raft leader for term 1. Latest log index is 0.
-2016-03-17T00:01:55.977866000 4221.937597s notice: Server got SIGINT from pid 518, uid 501; shutting down...
-2016-03-17T00:01:55.978213000 4221.937944s notice: Shutting down client connections...
-2016-03-17T00:01:55.978392000 4221.938123s notice: All client connections closed.
-2016-03-17T00:01:55.978398000 4221.938128s notice: Shutting down storage engine... (This may take a while if you had a lot of unflushed data in the writeback cache.)
-2016-03-17T00:01:56.079685000 4222.039416s notice: Storage engine shut down.
